src/services/runes.performance.service.ts
import { RPCClient } from '../utils/rpc.client';
import { RunesValidator } from '../utils/runes.validator';
import { Logger } from '../utils/logger';

export class RunesPerformanceService {
  constructor(
    private readonly rpcClient: RPCClient,
    private readonly validator: RunesValidator,
    private readonly logger: Logger
  ) {}

  async getMarketStats(): Promise<any> {
    try {
      // Implement market stats logic
      return {};
    } catch (error) {
      throw new Error('Failed to get market stats');
    }
  }

  async getRuneRanking(metric: string, limit: number): Promise<any> {
    try {
      // Implement rune ranking logic
      return [];
    } catch (error) {
      throw new Error('Failed to get ranking');
    }
  }
}
